requirements.txt
cvzone
opencv-python
numpy
mediapipe
PyMuPDF
streamlit
python-pptx